+++ src/test/java/se/bjurr/prnfb/service/UserCheckServiceTest.java
@@ -14,6 +14,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
 
+import se.bjurr.prnfb.presentation.dto.ON_OR_OFF;
 import se.bjurr.prnfb.settings.PrnfbButton;
 
 import com.atlassian.bitbucket.permission.PermissionService;
@@ -63,9 +64,9 @@ public class UserCheckServiceTest {
   when(this.userManager.isAdmin(this.userKey))//
     .thenReturn(false);
 
-  PrnfbButton button1 = new PrnfbButton(null, "title1", ADMIN, "off", "p1", "r1");
-  PrnfbButton button2 = new PrnfbButton(null, "title2", EVERYONE, "off", "p1", "r1");
-  PrnfbButton button3 = new PrnfbButton(null, "title3", SYSTEM_ADMIN, "off", "p1", "r1");
+  PrnfbButton button1 = new PrnfbButton(null, "title1", ADMIN, ON_OR_OFF.off, "p1", "r1");
+  PrnfbButton button2 = new PrnfbButton(null, "title2", EVERYONE, ON_OR_OFF.off, "p1", "r1");
+  PrnfbButton button3 = new PrnfbButton(null, "title3", SYSTEM_ADMIN, ON_OR_OFF.off, "p1", "r1");
   List<PrnfbButton> buttons = newArrayList(button1, button2, button3);
 
   Iterable<PrnfbButton> onlyAllowed = this.sut.filterAllowed(buttons);
@@ -82,7 +83,7 @@ public class UserCheckServiceTest {
     .thenReturn(this.userKey);
   when(this.userManager.isSystemAdmin(this.userKey))//
     .thenReturn(true);
-  PrnfbButton candidate = new PrnfbButton(null, "title", ADMIN, "off", "p1", "r1");
+  PrnfbButton candidate = new PrnfbButton(null, "title", ADMIN, ON_OR_OFF.off, "p1", "r1");
   assertThat(this.sut.isAllowedUseButton(candidate))//
     .isTrue();